+++ src/main/java/com/rapidtech/restapi/controller/ProductController.java
@@ -21,8 +21,7 @@ public class ProductController {
         this.service = service;
     }
 
-    @GetMapping
-
+        @GetMapping
         public ResponseEntity<Object> get(){
             List<ProductModel> result = service.getAll();
             return ResponseEntity.ok().body(
@@ -31,7 +30,7 @@ public class ProductController {
         }
 
         @GetMapping("/{id}")
-        public ResponseEntity<Object> getById(@PathVariable("id") Integer id){
+        public ResponseEntity<Object> getById(@PathVariable("id") Long id){
             Optional<ProductModel> result = service.getById(id);
             return ResponseEntity.ok().body(
                     new ResponseModel(200,"SUCCESS", result)
@@ -47,7 +46,7 @@ public class ProductController {
         }
 
         @PatchMapping("/{id}")
-        public ResponseEntity<Object> updateProduct(@PathVariable("id") Integer id, @RequestBody ProductModel request){
+        public ResponseEntity<Object> updateProduct(@PathVariable("id") Long id, @RequestBody ProductModel request){
             Optional<ProductModel> result = service.update(id, request);
             return ResponseEntity.ok().body(
                     new ResponseModel(200,"SUCCESS", result)
@@ -55,7 +54,7 @@ public class ProductController {
         }
 
         @DeleteMapping("/{id}")
-        public ResponseEntity<Object> delete(@PathVariable("id") Integer id){
+        public ResponseEntity<Object> delete(@PathVariable("id") Long id){
             Optional<ProductModel> result = service.delete(id);
             return ResponseEntity.ok().body(
                     new ResponseModel(200,"SUCCESS", result)
